GET spots updated with lat and lng changed from DECIMAL to FLOAT

+++ backend/db/migrations/20241229024949-create-spots.js
@@ -40,11 +40,15 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.DECIMAL(11, 8),
+        type: Sequelize.FLOAT,
+
+        // type: Sequelize.DECIMAL(11, 8),
         allowNull: false
       },
       lng: {
-        type: Sequelize.DECIMAL(11, 8),
+        type: Sequelize.FLOAT,
+
+        // type: Sequelize.DECIMAL(11, 8),
         allowNull: false
       },
       name: {
